dev server prepped for release
