+++ config.yml.example
@@ -2,14 +2,7 @@ development:
   password:
   theme_id: ""
   store:
-  directory: dist/
-  ignore_files:
-    - config/settings_data.json
-
-server:
-  password:
-  theme_id: ""
-  store:
+  storefront_api_key:
   directory: dist/
   ignore_files:
     - config/settings_data.json
