+++ webpack.dev.js
@@ -1,47 +1,83 @@
-const {
-    merge
-} = require('webpack-merge');
+const { merge } = require('we' + 'bpack-merge');
+const glob = require('glob');
 const common = require('./webpack.common.js');
 const path = require('path');
 const CopyPlugin = require('copy-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const WebpackShellPluginNext = require('webpack-shell-plugin-next');
+const { transformLiquid } = require('./shopify-dev-utils/transformLiquid');
+
+const port = 9000;
+const publicPath = `https://localhost:${port}/`;
 
 module.exports = merge(common, {
     mode: 'development',
     devtool: 'inline-source-map',
+    entry: glob.sync('./src/components/**/*.js').reduce(
+        (acc, path) => {
+            const entry = path.replace(/^.*[\\\/]/, '').replace('.js', '');
+            acc[entry] = path;
+            return acc;
+        },
+        { liquidDev: './shopify-dev-utils/liquidDev.entry.js' }
+    ),
     output: {
-        filename: './assets/bundle.[name].js',
+        filename: 'assets/bundle.[name].js',
+        hotUpdateChunkFilename: 'hot/[id].[fullhash].hot-update.js',
+        hotUpdateMainFilename: 'hot/[fullhash].hot-update.json',
         path: path.resolve(__dirname, 'dist'),
+        publicPath,
+    },
+    cache: false,
+    optimization: {
+        runtimeChunk: { name: 'runtime' },
     },
     module: {
-        rules: [{
-            test: /\.(sc|sa|c)ss$/,
-            use: [
-                MiniCssExtractPlugin.loader,
-                {
-                    loader: 'css-loader',
-                    options: {
-                        url: false,
+        rules: [
+            {
+                test: /\.liquid$/,
+                use: [
+                    { loader: 'raw-loader', options: { esModule: false } },
+                    {
+                        loader: path.resolve(__dirname, 'shopify-dev-utils/liquidDev.loader.js'),
+                        options: {
+                            publicPath,
+                            isSection(liquidPath) {
+                                const diff = path.relative(
+                                    path.join(__dirname, './src/components/'),
+                                    liquidPath
+                                );
+                                const componentType = diff.split(path.sep).shift();
+                                return componentType === 'sections';
+                            },
+                        },
                     },
-                },
-                'postcss-loader',
-                {
-                    loader: 'sass-loader',
-                    options: {
-                        sourceMap: true,
+                ],
+            },
+            {
+                test: /\.(sc|sa|c)ss$/,
+                use: [
+                    'style-loader',
+                    {
+                        loader: 'css-loader',
+                        options: {
+                            url: false,
+                        },
                     },
-                },
-            ],
-        }, ],
+                    'postcss-loader',
+                    {
+                        loader: 'sass-loader',
+                        options: {
+                            sourceMap: true,
+                        },
+                    },
+                ],
+            },
+        ],
     },
     plugins: [
         new WebpackShellPluginNext({
             onBuildStart: {
-                scripts: [
-                    'echo -- Webpack build started 🛠',
-                    'shopify-themekit watch --env=development',
-                ],
+                scripts: ['echo -- Webpack build started 🛠', 'shopify-themekit watch --env=development'],
                 blocking: false,
                 parallel: true,
             },
@@ -76,7 +112,7 @@ module.exports = merge(common, {
                         const targetFolder = diff.split(path.sep)[0];
                         return path.join(targetFolder, path.basename(absolutePath));
                     },
-                    transform: undefined,
+                    transform: transformLiquid(publicPath),
                 },
                 {
                     from: 'src/assets/**/*',
@@ -93,5 +129,21 @@ module.exports = merge(common, {
                 },
             ],
         }),
-    ].filter(Boolean),
+    ],
+    devServer: {
+        contentBase: path.join(__dirname, 'dist'),
+        publicPath: '/',
+        headers: {
+            'Access-Control-Allow-Origin': '*',
+            'Access-Control-Allow-Methods': 'GET, POST, PUT, DELETE, PATCH, OPTIONS',
+            'Access-Control-Allow-Headers': 'X-Requested-With, content-type, Authorization',
+        },
+        compress: true,
+        port,
+        https: true,
+        disableHostCheck: true,
+        hot: true,
+        overlay: true,
+        writeToDisk: true,
+    },
 });
